pkg/scheduler: Abstract task scheduling
This commit re-organises the way in which tasks are run such that
alternative ways can be introduced easily based on different needs.
The current scheduler, "simple", places tasks (based on core count)
onto a map of cores, such that the `run` has exactly the cores it
needs and nothing else.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>